message apis clean
