+++ src/controllers/messagingController.js
@@ -1,18 +1,6 @@
 import MessagingClass from "../classes/messagingClass.js";
 
 const messagingController = {
-  listTutors: async (req, res) => {
-    const studentId = req.user._id;
-    const data = await new MessagingClass().listTutorsForStudents(studentId);
-    res.status(data.statusCode).json(data);
-  },
-
-  listStudents: async (req, res) => {
-    const tutorId = req.user._id;
-    const data = await new MessagingClass().listStudentsForTutor(tutorId);
-    res.status(data.statusCode).json(data);
-  },
-
   sendMessage: async (req, res) => {
     const { chatId, messageContent, receiverType, receiverId } = req.body;
